feat(engine): extract path-matchers.ts's pure classifier family to gittensory-engine (#4252)
